removed lots of old unmerged apps
many of these apps are still being developed or used in the rapidsms
trunk, but haven't been updated for a long time in my (or ewheeler's)
fork, so are no longer compatible with my lib/rapidsms. i'm not going
to keep _everything_ working as i go (especially since so many of my
grand API changes turn out not to be worth the upset, and are quietly
reverted), so i'm deleting the apps which i'm ignoring, to make it
very clear which ones will need to be fixed manually during mfchang's
upcoming MERGEAPALOOZA.

hopefully soon, ALL non-core/contrib apps will be deleted from this
repo, and moved (along with their history) to their own repos.